General improvements
* Added automated error output
* Error handling for unexpected success
* Using cargo build now instead of cargo run
* Added max_tries variable
* Fixed skipping + annotation logic
* Added ability to create folder

+++ scripts/update_tests.sh
@@ -32,7 +32,6 @@ $0 -r . -s                      # For all current tests, check if @unittest.skip
     * The updater does not add skips to classes, only on tests
     * If there are multiple tests with the same name, a decorator will be added to all of them
     * The updater does not retain anything more specific than a general skip (skipIf/Unless will be replaced by a general skip)
-    * Currently, the updater does not take unexpected successes into account
 
 EOF
     exit 1
@@ -52,6 +51,7 @@ check_skip_flag=false
 num_jobs=20
 libraries=()
 ignored_libraries=("multiprocessing" "concurrent")
+max_tries=20
 
 while [[ $# -gt 0 ]]; do
     case "$1" in
@@ -93,6 +93,8 @@ while [[ $# -gt 0 ]]; do
             ;;
     esac
 done
+
+cargo build --release --features encodings,sqlite 
 # -------------------------------------- Constants ------------------------------------- #
 RUSTPYTHON_POSSIBLE_SKIP_RE="@unittest.skip.*\([\"']TODO:\s*RUSTPYTHON.*[\"']\)"
 RUSTPYTHON_CANONICAL_SKIP_RE="@unittest.skip\('TODO: RUSTPYTHON; .*'\)"
@@ -116,15 +118,11 @@ update_test() {
     local clib_path="$cpython_path/$lib"
     local rlib_path="$rpython_path/$lib"
 
-    if [[ -f "$clib_path" && -f "$rlib_path" ]] && files_equal "$clib_path" "$rlib_path"; then
-        echo "No changes in $lib. Skipping..." 
-        return
-    fi
-
-    if [[ ! -f "$rlib_path" ]]; then
+    if [[ ! -f "$rlib_path" ]] && ! [[ -f "$clib_path" && -f "$rlib_path" ]] && ! files_equal "$clib_path" "$rlib_path"; then
         echo "Test file $lib missing"
         if $copy_untracked; then
             echo "Copying $lib ..."
+            mkdir -p "$(dirname "$rlib_path")"
             cp "$clib_path" "$rlib_path"
         fi
     else
@@ -133,7 +131,7 @@ update_test() {
     fi
 
 
-    if [[ $annotate && -f "$rlib_path" && $(basename -- "$rlib_path") == test_*.py ]]; then
+    if $annotate && [[ -f "$rlib_path" && $(basename -- "$rlib_path") == test_*.py ]]; then
         annotate_lib $lib $rlib_path
     fi
 }
@@ -214,15 +212,18 @@ annotate_lib() {
     while ! grep -q "Tests result: SUCCESS" <<< "$output"; do
         ((attempts++))
         # echo "$lib failing, annotating..."
-        readarray -t failed_tests <<< "$(echo "$output" | awk '/^(FAIL:|ERROR:)/ {print $2}' | sort -u)"
-        
+        readarray -t failed_tests <<< "$(echo "$output" | awk '/^(FAIL:|ERROR:|UNEXPECTED SUCCESS:)/ {print $(NF-1)}' | sort -u)"
+
         # If the test fails/errors, then expectedFailure it
         for test in "${failed_tests[@]}"
         do
-            if already_failed $rlib_path $test; then
-                replace_expected_with_skip $rlib_path $test
+            error=$(echo "$output" | pcregrep -M -o2 "(?s)======================================================================\n\s*(FAIL:|ERROR:) $test.*?\n----------------------------------------------------------------------.*\n(.*)\n\n(======================================================================|----------------------------------------------------------------------)" 2>/dev/null)
+            if echo "$output" | grep -q "UNEXPECTED SUCCESS: $test"; then
+                replace_expected_with_skip $rlib_path $test "Unexpected success $error"
+            elif already_failed $rlib_path $test; then
+                replace_expected_with_skip $rlib_path $test "$error"
             else
-                add_above_test $rlib_path $test "$RUSTPYTHON_CANONICAL_EX_FAILURE" 
+                add_above_test $rlib_path $test "$RUSTPYTHON_CANONICAL_EX_FAILURE; $error"
             fi
         done
 
@@ -239,7 +240,7 @@ annotate_lib() {
 
         output=$(rustpython $lib 2>&1)
 
-        if [[ $attempts -gt 15 ]]; then 
+        if [[ $attempts -gt $max_tries ]]; then 
             echo "Issue annotating $lib" >&2
             return;
         fi
@@ -265,13 +266,14 @@ add_above_test() {
 # --------------------------------- Utility functions --------------------------------- #
 
 rustpython() {
-    cargo run --release --features encodings,sqlite -- -m test -j 1 -u all --fail-env-changed --timeout "$timeout" -v "$@"
+    target/release/rustpython -m test -j 1 -u all --fail-env-changed --timeout "$timeout" -v "$@"
 }
 
 replace_expected_with_skip() {
     file=$1
     test_name=$2
-    sed -E "/$RUSTPYTHON_CANONICAL_EX_FAILURE_RE/ { N; /\n\s*def $test_name/ { s/^(\s*)@unittest\.expectedFailure\s+# TODO: RUSTPYTHON/\1@unittest.skip\('TODO: RUSTPYTHON'\)/ } }" -i $file
+    error=$3
+    sed -E "/$RUSTPYTHON_CANONICAL_EX_FAILURE_RE/ { N; /\n\s*def $test_name/ { s/^(\s*)@unittest\.expectedFailure\s+# TODO: RUSTPYTHON.*\n/\1@unittest.skip\('TODO: RUSTPYTHON; $error'\)\n/ } }" -i $file
 }
 
 already_failed() {
